+++ yash-env/src/system/future.rs
@@ -34,6 +34,7 @@ use std::task::{Context, Poll};
 /// checking to access its internal state guarded by a `RefCell`. Instead of
 /// borrowing the system, the future must share ownership of the system to keep
 /// it alive until the future is resolved.
+#[deprecated(note = "FlexFuture is no longer used in this crate", since = "0.12.0")]
 pub enum FlexFuture<T> {
     /// Future that is already ready with a value
     Ready(std::future::Ready<T>),
@@ -43,6 +44,7 @@ pub enum FlexFuture<T> {
     Generic(Pin<Box<dyn Future<Output = T>>>),
 }
 
+#[allow(deprecated)]
 impl<T: Debug> Debug for FlexFuture<T> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
@@ -53,36 +55,42 @@ impl<T: Debug> Debug for FlexFuture<T> {
     }
 }
 
+#[allow(deprecated)]
 impl<T> From<T> for FlexFuture<T> {
     fn from(value: T) -> Self {
         FlexFuture::Ready(std::future::ready(value))
     }
 }
 
+#[allow(deprecated)]
 impl<T> From<std::future::Ready<T>> for FlexFuture<T> {
     fn from(ready: std::future::Ready<T>) -> Self {
         FlexFuture::Ready(ready)
     }
 }
 
+#[allow(deprecated)]
 impl<T> From<std::future::Pending<T>> for FlexFuture<T> {
     fn from(pending: std::future::Pending<T>) -> Self {
         FlexFuture::Pending(pending)
     }
 }
 
+#[allow(deprecated)]
 impl<T> From<Pin<Box<dyn Future<Output = T>>>> for FlexFuture<T> {
     fn from(future: Pin<Box<dyn Future<Output = T>>>) -> Self {
         FlexFuture::Generic(future)
     }
 }
 
+#[allow(deprecated)]
 impl<T> From<Box<dyn Future<Output = T>>> for FlexFuture<T> {
     fn from(future: Box<dyn Future<Output = T>>) -> Self {
         FlexFuture::Generic(Box::into_pin(future))
     }
 }
 
+#[allow(deprecated)]
 impl<T> FlexFuture<T> {
     /// Creates a new `FlexFuture` from any future.
     ///
@@ -108,12 +116,14 @@ impl<T> FlexFuture<T> {
     }
 }
 
+#[allow(deprecated)]
 impl<T: 'static> From<FlexFuture<T>> for Pin<Box<dyn Future<Output = T>>> {
     fn from(future: FlexFuture<T>) -> Self {
         future.into_boxed()
     }
 }
 
+#[allow(deprecated)]
 impl<T> Future for FlexFuture<T> {
     type Output = T;